requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
flake8
# psycopg2>=2.8.6,<2.9
psycopg2-binary>=2.8.6,<2.9
